# src/main/java/io/umiverse/gateway/openapi/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.umiverse.gateway.openapi.api;

import io.umiverse.gateway.openapi.model.LoginRequest;
import io.umiverse.gateway.openapi.model.Response;
import io.umiverse.gateway.openapi.model.User;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-08-03T15:10:02.892036+08:00[Asia/Shanghai]")
@Validated
@Tag(name = "auth", description = "the auth API")
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /auth/apple/login : Login using apple service.
     *
     * @param loginRequest  (optional)
     * @return Login success, user info returned. (status code 200)
     *         or Login failed, invalid id token (status code 403)
     */
    @Operation(
        operationId = "authAppleLoginPost",
        summary = "Login using apple service.",
        tags = { "auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Login success, user info returned.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "403", description = "Login failed, invalid id token")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/apple/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> authAppleLoginPost(
        @Parameter(name = "LoginRequest", description = "") @Valid @RequestBody(required = false) LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : 6, \"subscriber\" : \"subscriber\", \"registerTime\" : 1, \"id\" : 0, \"userName\" : \"userName\", \"locale\" : \"locale\", \"email\" : \"email\", \"picture\" : \"picture\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/google/login : Login using google service.
     *
     * @param loginRequest  (optional)
     * @return Login success, user info returned. (status code 200)
     *         or Login failed, invalid id token (status code 403)
     */
    @Operation(
        operationId = "authGoogleLoginPost",
        summary = "Login using google service.",
        tags = { "auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Login success, user info returned.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "403", description = "Login failed, invalid id token")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/google/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> authGoogleLoginPost(
        @Parameter(name = "LoginRequest", description = "") @Valid @RequestBody(required = false) LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : 6, \"subscriber\" : \"subscriber\", \"registerTime\" : 1, \"id\" : 0, \"userName\" : \"userName\", \"locale\" : \"locale\", \"email\" : \"email\", \"picture\" : \"picture\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/login : Login using firebase apple/google/twitter/facebook... etc service.(web3auth)
     *
     * @param loginRequest  (optional)
     * @return Login success, user info returned. (status code 200)
     *         or Login failed, invalid id token (status code 403)
     */
    @Operation(
        operationId = "authLoginPost",
        summary = "Login using firebase apple/google/twitter/facebook... etc service.(web3auth)",
        tags = { "auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Login success, user info returned.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "403", description = "Login failed, invalid id token")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> authLoginPost(
        @Parameter(name = "LoginRequest", description = "") @Valid @RequestBody(required = false) LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : 6, \"subscriber\" : \"subscriber\", \"registerTime\" : 1, \"id\" : 0, \"userName\" : \"userName\", \"locale\" : \"locale\", \"email\" : \"email\", \"picture\" : \"picture\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /auth/logout : user logout.
     *
     * @return user logout success. (status code 200)
     *         or  (status code 401)
     *         or Logout failed, invalid auth token (status code 403)
     */
    @Operation(
        operationId = "authLogoutGet",
        summary = "user logout.",
        tags = { "auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "user logout success.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Response.class))
            }),
            @ApiResponse(responseCode = "401", description = ""),
            @ApiResponse(responseCode = "403", description = "Logout failed, invalid auth token")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/auth/logout",
        produces = { "application/json" }
    )
    default ResponseEntity<Response> authLogoutGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : \"{}\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}