# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: API definition for Umiverse server apis
  title: Umiverse Server API Spec
  version: 2.0.0
servers:
- url: https://api-dev.umiverse.io/
...
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: "Login success, user info returned."
        "403":
          description: "Login failed, invalid id token"
      summary: Login using firebase apple/google/twitter/facebook... etc service.(web3auth)
      tags:
      - auth
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: auth
  /auth/google/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: "Login success, user info returned."
        "403":
          description: "Login failed, invalid id token"
      summary: Login using google service.
      tags:
      - auth
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: auth
  /auth/apple/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: "Login success, user info returned."
        "403":
          description: "Login failed, invalid id token"
      summary: Login using apple service.
      tags:
      - auth
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: auth
  /auth/logout:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
          description: user logout success.
        "401":
          $ref: '#/components/schemas/UnauthorizedError'
        "403":
          description: "Logout failed, invalid auth token"
      security:
      - bearerAuth: []
      summary: user logout.
      tags:
      - auth
      x-accepts: application/json
      x-tags:
      - tag: auth
...
  /game/list:
    get:
      description: This api returns all games as list.
      parameters:
      - description: "GameList order, ALL or NEW, default is ALL"
        explode: true
        in: query
        name: order
        required: false
        schema:
          type: string
        style: form
      - description: Choose game by category
        explode: true
        in: query
        name: category
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Game'
                maxItems: 100
                type: array
          description: Get game list success.
      summary: Get all game list.
      tags:
      - game
      x-accepts: application/json
      x-tags:
      - tag: game
  /game/{gameId}:
    get:
      parameters:
      - description: the request game ID to search.
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
          description: Get game info success.
      summary: Get game detail from server
      tags:
      - game
      x-accepts: application/json
      x-tags:
      - tag: game
  /game/{gameId}/items:
    get:
      parameters:
      - description: the request game ID to search.
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GameItem'
                type: array
          description: Get game item list success.
      summary: Get game items by gameId from server
      tags:
      - game
      x-accepts: application/json
      x-tags:
      - tag: game
...
components:
  schemas:
    UnauthorizedError:
      description: Access token is missing or invalid
    Response:
      example:
        data: "{}"
        message: message
      properties:
        data:
          description: return data of response
          type: object
        message:
          description: message of response
          type: string
      type: object
    LoginRequest:
      example:
        jwtToken: jwtToken
      properties:
        jwtToken:
          description: jwt token fetched from firebase.
          type: string
      required:
      - jwtToken
      type: object
    User:
      example:
        emailVerified: 6
        subscriber: subscriber
        registerTime: 1
        id: 0
        userName: userName
        locale: locale
        email: email
        picture: picture
      properties:
        id:
          description: userId saved in system
          format: int64
          type: integer
        userName:
          description: name of the user
          type: string
        email:
          description: email address of the user
          type: string
        emailVerified:
          description: Is the email verified(1)
          type: integer
        picture:
          description: image of the user
          type: string
        locale:
          description: locale of the user
          type: string
        subscriber:
          description: Unique key for the user
          type: string
        registerTime:
          description: Time stamp of user created time.
          format: int64
          type: integer
      type: object
...
    Game:
      example:
        coverImageUrl: coverImageUrl
        line: line
        nid: nid
        description: description
        orderWeight: 5
        screenshot: screenshot
        secret: secret
        createTimestamp: 5
        faviconUrl: faviconUrl
        twitter: twitter
        gameServerUrl: gameServerUrl
        company: company
        callbackUrl: callbackUrl
        id: 0
        iconUrl: iconUrl
        state: 6
        introduction: introduction
        email: email
        highlightType: 1
        cdnUrl: cdnUrl
        facebook: facebook
        name: name
        shortName: shortName
        category: category
        slogan: slogan
      properties:
        id:
          description: GameId saved in system
          format: int64
          type: integer
        nid:
          description: 游戏版本号
          type: string
        name:
          description: 游戏名
          type: string
        shortName:
          description: 游戏短名
          type: string
        category:
          description: 游戏分类
          type: string
        slogan:
          description: 标题下短句
          type: string
        description:
          description: 中篇描述
          type: string
        introduction:
          description: 详细页用长篇介绍
          type: string
        iconUrl:
          description: icon
          type: string
        faviconUrl:
          description: small favicon
          type: string
        screenshot:
          description: list of 5 pics
          type: string
        coverImageUrl:
          description: 封面图片
          type: string
        email:
          description: 游戏CP公司邮箱
          type: string
        company:
          description: Company name of the game.
          type: string
        twitter:
          description: Company twitter.
          type: string
        facebook:
          description: facebook
          type: string
        line:
          description: line
          type: string
        cdnUrl:
          description: Cdn download url.
          type: string
        gameServerUrl:
          description: 游戏服务器Url
          type: string
        callbackUrl:
          description: Call back url.
          type: string
        state:
          description: 游戏状态 0 准备;1 测试;2 公开
          type: integer
        highlightType:
          description: "列表中的高亮,Normal\\New\\Hot"
          type: integer
        orderWeight:
          description: 游戏列表中排列顺序
          type: integer
        secret:
          description: "游戏验证用secret,与id并用"
          type: string
        createTimestamp:
          description: created time.
          format: int64
          type: integer
      type: object
    GameItem:
      example:
        gameId: 6
        price: 1
        pictureUrl: pictureUrl
        name: name
        id: 0
      properties:
        id:
          description: Id of the game item.
          format: int64
          type: integer
        gameId:
          description: Id of the game.
          format: int64
          type: integer
        name:
          description: name of the item
          type: string
        pictureUrl:
          description: url of the picture
          type: string
        price:
          description: price of this item.
          format: int64
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http